Add files via upload
Hello everyone. This is a mini project made my me as an intern in InTernPe under the domain of web development. This is a simple calculator made using some basic concepts of HTML, CSS & JavaScript. It was really fun spending time on this project. It made me aware of the front-end development as well.